requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cssselect==1.3.0
exceptiongroup==1.2.2
feedfinder2==0.0.4
feedparser==5.2.1
feedsearch==1.0.12
h11==0.16.0
idna==3.10
jieba3k==0.35.1
joblib==1.5.0
lxml==5.4.0
lxml-html-clean==0.4.2
MarkupSafe==3.0.2
newscatcher==0.2.0
newspaper3k==0.2.8
nltk==3.9.1
outcome==1.3.0.post0
pillow==11.2.1
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
selenium==4.32.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
tinysegmenter==0.3
tldextract==2.2.3
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==2.2.2
wsproto==1.2.0
